Creada la entidad de voto, y su respectivo repositorio, además de empezar con la lógica de la popularidad de cada oferta

+++ backend/src/main/java/com/gym360/backend/repository/UsuarioRepositorio.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface UsuarioRepositorio extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepositorio extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByNombre(String username);
 
     Usuario findByEmail(String email);
